Jay

See merge request COMP2511/22T2/teams/T15A_ECLAIR/project!33

+++ src/main/java/dungeonmania/player/NormalState.java
@@ -14,7 +14,7 @@ public class NormalState implements PlayerState {
         return "Normal";
     }
 
-    public int getAttack() {
+    public double getAttack() {
         return player.attack;
     }
